+++ client/components/index.js
@@ -5,7 +5,6 @@
  */
 export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
-export {Login, Signup} from './auth-form'
 export {
   default as AllInstrumentsFromFirestore
 } from './AllInstrumentsWithFirestore'
