+++ client/components/user-home.js
@@ -1,30 +1,36 @@
 import React, {Component} from 'react'
 import {db, auth} from '../firestore/db'
+import Modal from 'react-bootstrap/Modal'
+import Button from 'react-bootstrap/Button'
 
 export default class UserHome extends Component {
   constructor() {
     super()
-    this.state = {list: []}
+    this.state = {
+      list: []
+    }
   }
+
   async componentDidMount() {
     const docRef = await db
       .collection('users')
       .doc(auth.currentUser.uid)
       .get()
 
-    const projects = docRef.data().projects
-    console.log('CODYS project', projects)
+    let projects
     let projectsBody
-
-    if (projects.id) {
-      const projectsPromise = await db
-        .collection('projects')
-        .doc(projects.id)
-        .get()
-      if (projectsPromise) {
-        projectsBody = projectsPromise.data()
-      } else {
-        projectsBody = null
+    if (docRef.data().projects) {
+      projects = docRef.data().projects
+      if (projects.id) {
+        const projectsPromise = await db
+          .collection('projects')
+          .doc(projects.id)
+          .get()
+        if (projectsPromise) {
+          projectsBody = projectsPromise.data()
+        } else {
+          projectsBody = null
+        }
       }
     } else {
       projectsBody = null
@@ -36,6 +42,7 @@ export default class UserHome extends Component {
   render() {
     console.log(this.state)
     const email = auth.currentUser.email
+
     return (
       <div>
         <h3>Welcome, {email}</h3>
