+++ client/routes.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {withRouter, Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
-import {Login, Signup, UserHome} from './components'
+import {UserHome} from './components'
 import {me} from './store'
 import AllInstruments from './components/AllInstruments'
 import {auth} from './firestore/db'
@@ -11,15 +11,12 @@ class Routes extends Component {
   render() {
     return (
       <Switch>
-        <Route path="/login" component={Login} />
-        <Route path="/signup" component={Signup} />
         <Route path="/play" component={AllInstruments} />
         {auth.currentUser && (
           <Switch>
             <Route path="/home" component={UserHome} />
           </Switch>
         )}
-        <Route component={Login} />
       </Switch>
     )
   }
